separate graphql and rest api into two classes
